span-tag: allow to deduplicate against SOLR
If -server is given (to a VuFind/SOLR), try to run clean ISIL
attachments by comparing DOI with documents in index.

A hacky way to drop labels by querying a SOLR server for DOI (we do not
have a dedicated field yet, so query 'allfields', this can lead to
problems with prefixes, too) and then compare source ids in index with
preferences, adjustable via -prefs, similar to groupcover(1).

This obviously takes longer, as we run one query for each DOI. The query
is not yet optimized. To speed up processing, increase -w number of
worker (and lower -b batch size, to not run out of memory).